Getting Started - Tutorial 6/TestNinja/TestNinja/Fundamentals/Math.cs
using System.Collections.Generic;

namespace TestNinja.Fundamentals
{
    public class Math
    {
        public int Add(int a, int b)
        {
            return a + b;
        }
        
        public int Max(int a, int b)
        {
            return (a > b) ? a : b;
        }

        //Section 3 - Tutorial 3 - Testing Arrays
        //This Function Returns an Array of Odd Number.
        public IEnumerable<int> GetOddNumbers(int limit)
        {
            for (var i = 0; i <= limit; i++)
                if (i % 2 != 0)
                    yield return i; 
        }
    }
}